Primer Commit

+++ app/Http/Controllers/RoleController.php
@@ -3,10 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Interfaces\MainControllerInterface;
-use App\Models\Business;
 use App\Models\NewRole;
 use App\Models\User;
-use App\Repositories\BusinessRepository;
 use App\Repositories\LogRepository;
 use App\Repositories\RoleRepository;
 use Illuminate\Contracts\Routing\ResponseFactory;
@@ -16,8 +14,10 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
-use \Illuminate\Contracts\Foundation;
+use Illuminate\Contracts\Foundation;
+use Illuminate\Validation\Rule;
 use Exception;
+use Throwable;
 
 
 /**
@@ -57,8 +57,7 @@ class RoleController extends BaseController implements MainControllerInterface {
         $rq = getRequestParams($request);
         $roles = $this->roleRepository->search($rq->search);
         if(!auth()->user()->hasAnyRole(ALL_ACCESS)){
-            $roles = $roles->whereIn('name', auth()->user()->getRoleNames()->toArray())
-                ->whereHas(BUSINESS_IDENTIFY);
+            $roles = $roles->whereIn('name', auth()->user()->getRoleNames()->toArray());
         }
         $roles = $roles->sortable();
         return  $this->roleRepository->getResponse($roles, $rq);
@@ -70,16 +69,25 @@ class RoleController extends BaseController implements MainControllerInterface {
      * @throws Exception
      */
     public function store(Request $request): Response|Foundation\Application|ResponseFactory {
-        $request['public'] = $request['public'] == 'on';
-        $request['is_admin'] = $request['is_admin'] == 'on';
-        $input = $request->all();
-
         try {
+            $request->validate([
+                'name' => [
+                    'required',
+                    Rule::unique('roles')->whereNull('deleted_at'),
+                ],
+            ],[
+                'name.unique' => __('The [name] field must be unique, a record with this name already exists'),
+            ]);
+
+            $input = $request->all();
             DB::beginTransaction();
             $this->roleRepository->create($input);
+
+            auth()->user()->assignRole($input['name']);
+
             DB::commit();
             return response(__('Success'), 200);
-        }catch (\Throwable $e){
+        }catch (Throwable $e){
             DB::rollBack();
             return response($e->getMessage(), 403);
         }
@@ -93,8 +101,6 @@ class RoleController extends BaseController implements MainControllerInterface {
         $this->setIcons();
         $obj    =   new $this->object($this->getParams($request, false));
         return response()->json([
-            'rating_types_url' => route('comment.rating_types', app()->getLocale()),
-            'comments' => route('comment.all', app()->getLocale()),
             'object' => $this->object,
             'title' => __(strtoupper($obj->singular)),
             'csrf' => csrf_token(),
@@ -116,17 +122,18 @@ class RoleController extends BaseController implements MainControllerInterface {
      * @param String $lang
      * @param int $id
      * @return Foundation\Application|ResponseFactory|JsonResponse|Response
+     * @throws Exception
      */
     public function show(String $lang, int $id): Response|JsonResponse|Foundation\Application|ResponseFactory {
         $fields = $this->roleRepository->makeModel()->getFields();
         $role =  $this->roleRepository->find($id);
         if (empty($role)) return response(__('Not found'), 404);
         return response()->json([
-            'object' => $this->roleRepository->makeModel()->whereHas(BUSINESS_IDENTIFY)->with($this->roleRepository->includes)->where('id' ,$id)->first(),
+            'object' => $this->roleRepository->makeModel()->with($this->roleRepository->includes)->where('id' ,$id)->first(),
             'fields' => $fields,
             'icons' => getAllIcons(),
             'csrf' => csrf_token(),
-            'title'=> __('Show the role'),
+            'title'=> 'role',
             'url' => '#'
         ]);
     }
@@ -135,17 +142,18 @@ class RoleController extends BaseController implements MainControllerInterface {
      * @param String $lang
      * @param int $id
      * @return Foundation\Application|ResponseFactory|JsonResponse|Response
+     * @throws Exception
      */
     public function edit(String $lang, int $id): Response|JsonResponse|Foundation\Application|ResponseFactory {
         $fields = $this->roleRepository->makeModel()->getFields();
         $role =  $this->roleRepository->find($id);
         if (empty($role)) return response(__('Not found'), 404);
         return response()->json([
-            'object' => $this->roleRepository->makeModel()->whereHas(BUSINESS_IDENTIFY)->with($this->roleRepository->includes)->where('id', $id)->first(),
+            'object' => $this->roleRepository->makeModel()->with($this->roleRepository->includes)->where('id', $id)->first(),
             'fields' => $fields,
             'icons' => getAllIcons(),
             'csrf' => csrf_token(),
-            'title'=> __('Update the role'),
+            'title'=> 'role',
             'url' => route('role.update', ['locale' => $lang, 'role' => $id])
         ]);
     }
@@ -158,16 +166,23 @@ class RoleController extends BaseController implements MainControllerInterface {
      * @throws Exception
      */
     public function update(Request $request, String $lang, int $id): Response|Foundation\Application|ResponseFactory {
-        $request['public'] = $request['public'] == 'on';
-        $request['is_admin'] = $request['is_admin'] == 'on';
-        $input = $request->all();
-        $role = $this->roleRepository->find($id);
         try {
+            $request->validate([
+                'name' => [
+                    'required',
+                    Rule::unique('roles')->whereNull('deleted_at')->ignore($id),
+                ],
+            ],[
+                'name.unique' => __('The [name] field must be unique, a record with this name already exists'),
+            ]);
+
+            $input = $request->all();
+            $role = $this->roleRepository->find($id);
             DB::beginTransaction();
             $role->update($input);
             DB::commit();
             return response(__('Success'), 200);
-        }catch (\Throwable $e){
+        }catch (Throwable $e){
             DB::rollBack();
             return response($e->getMessage(), 403);
         }
@@ -182,19 +197,19 @@ class RoleController extends BaseController implements MainControllerInterface {
     public function destroy(String $lang, int $id): Response|JsonResponse|Foundation\Application|ResponseFactory {
         $role =  $this->roleRepository->find($id);
         if (empty($role)) return response(__('Not found'), 404);
-        if($role->name == 'Developer') throw new Exception(__("Not all roles should be removed"));
         try {
+            if($role->name == 'Developer') throw new Exception(__("Not all roles should be removed"));
+            if($role->name == 'General') throw new Exception(__("The role named General is necessary"));
             DB::beginTransaction();
             $users = User::all();
             foreach($users as $user){
                 if( $user->hasRole($role->name))  $user->removeRole($role->name);
             }
             $role->permissions()->detach();
-            $role->business()->detach();
             $role->delete();
             DB::commit();
             return response()->json(['delete' => 'success']);
-        }catch (\Throwable $e){
+        }catch (Throwable $e){
             DB::rollBack();
             return response($e->getMessage(), 403);
         }
@@ -207,18 +222,5 @@ class RoleController extends BaseController implements MainControllerInterface {
      */
     public function logs(Request $request, string $lang): Response|JsonResponse {
         return getAllModelLogs($request,NewRole::class, $this->logRepository);
-    }
-
-    /**
-     * @param Request $request
-     * @return Factory|View|Foundation\Application
-     */
-    public function assignBusiness(Request $request): Factory|View|Foundation\Application {
-        return view('pages.assignments.assign', [
-            'url' => route('assign.objects', app()->getLocale()) ,
-            'rows' => Business::class,
-            'columns' => NewRole::class,
-            'key' => 'role_for_business'
-        ]);
     }
 }
